+++ packages/client/src/pages/authentication/Logout.jsx
