+++ packages/client/src/pages/authentication/auth.guard.jsx
