+++ packages/client/src/pages/course/CourseDetailGuard.jsx
@@ -1,6 +1,6 @@
 import { useAuth } from '../../context/auth.context.jsx';
-import { Students } from './professor/courseDetail.jsx';
-import { CourseDetail } from './student/courseDetail.jsx';
+import { Students } from './professor/CourseDetail.jsx';
+import { CourseDetail } from './student/CourseDetail.jsx';
 
 export const CourseDetailGuard = () => {
   const { decoded_token } = useAuth();
