+++ packages/client/src/pages/course/professor/CourseDetail.jsx
@@ -1,5 +1,5 @@
 import { Button, IconButton, Stack, Tooltip, Typography } from '@mui/material';
-import { StudentTable } from './studentTable.jsx';
+import { StudentTable } from './StudentTable.jsx';
 import { Delete } from '@mui/icons-material';
 import { useEffect, useState } from 'react';
 import { useAuth } from '../../../context/auth.context.jsx';
@@ -7,7 +7,7 @@ import { useLocation, useParams } from 'react-router-dom';
 import { axiosInstance } from '../../../utils/axioInstance.js';
 import { Paths } from '../../../constants/Paths.js';
 import { Link } from 'react-router-dom';
-import { DeleteStudentDialog } from './deleteStudentDialog.jsx';
+import { DeleteStudentDialog } from './DeleteStudentDialog.jsx';
 import { useCourseDetail } from '../../../hooks/useCourseDetail.jsx';
 
 export const Students = () => {
