+++ packages/client/src/pages/course/professor/DeleteStudentDialog.jsx
