+++ packages/client/src/pages/home/Layout.jsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
 import { Container } from '@mui/material';
 import { Navbar } from './component/navbar.jsx';
-import { SplitScreen } from './component/splitScreen.jsx';
+import { SplitScreen } from './component/SplitScreen.jsx';
 
 export const Layout = () => {
   return (
