+++ packages/client/src/pages/home/component/DeleteCourseDialog.jsx
