+++ packages/client/src/pages/home/component/SplitScreen.jsx
